Clarify application module ownership
Move lifecycle and formatting code to their actual owners, require explicit LLM system prompts, and import emulator records from their parser modules. Group rolling-memory schemas, prompts, and persistence/compaction workflow in one feature package to remove the dependency cycle.
