+++ agent/battle/agent.py
@@ -12,7 +12,7 @@ from agent.context import AgentContext
 from agent.utils import AGENT_HOOKS, build_screenshot_content, is_battle_handler_state
 from common.prompts import SYSTEM_PROMPT
 from llm.service import MODEL, REASONING_EFFORT, TIMEOUT_SECONDS
-from memory.rolling_memory import finalize_iteration
+from memory.rolling_memory.service import finalize_iteration
 
 if TYPE_CHECKING:
     from PIL import Image
