+++ agent/overworld/tools/navigate/service.py
@@ -12,7 +12,7 @@ if TYPE_CHECKING:
     from common.schemas import Coords
     from emulator.emulator import Emulator
     from emulator.game_state import GameState
-    from memory.rolling_memory import RollingMemory
+    from memory.rolling_memory.schemas import RollingMemory
     from overworld_map.schemas import OverworldMap
