+++ agent/text/agent.py
@@ -17,7 +17,7 @@ from agent.text.utils import (
 from agent.utils import AGENT_HOOKS, build_screenshot_content
 from common.prompts import SYSTEM_PROMPT
 from llm.service import MODEL, REASONING_EFFORT, TIMEOUT_SECONDS
-from memory.rolling_memory import finalize_iteration
+from memory.rolling_memory.service import finalize_iteration
 
 if TYPE_CHECKING:
     from PIL import Image
